Adjusted phone number parsing method

+++ ex01/Contact.hpp
@@ -11,7 +11,7 @@ class	Contact
 		std::string		first_name;
 		std::string		last_name;
 		std::string		nickname;
-		int				phone_number;
+		long long		phone_number;
 		std::string		darkest_secret;
 	
 	Contact() : index(-1) {};
